refactor: 영어 버전 수정
- 영어 버전 수정 및 추가했습니다.
