+++ gacha/gacha/Views/DashBoard/MainViewAfter.swift
@@ -110,7 +110,7 @@ struct MainViewAfter: View {
         HStack(spacing: 16) {
             // TODO: 통계 데이터 표시
             VStack(alignment: .leading, spacing: 4) {
-                Text("굴곡 각도")
+                Text(Strings.Progress.flexionAngle)
                     .font(.displayCalloutBold)
                     .foregroundColor(.blue700)
                 Text(vm.formatAngle(vm.currentRecord?.flexionAngle))
@@ -120,7 +120,7 @@ struct MainViewAfter: View {
             Spacer()
             
             VStack(alignment: .leading, spacing: 4) {
-                Text("통증 정도")
+                Text(Strings.Progress.painLevel)
                     .font(.displayCalloutBold)
                     .foregroundColor(.blue700)
                 Text("\(vm.currentRecord?.painLevel ?? 0)")
