+++ gacha/gacha/en.lproj/Localizable.strings
@@ -11,6 +11,7 @@
 "common.no" = "No";
 "common.retake" = "Retake Measurement";
 "common.cancel" = "Cancel";
+"common.confirm" = "Confirm";
 
 /* Button */
 "button.next" = "Next";
@@ -53,8 +54,9 @@
 "flexion.measured_semibold24" = "Completed";
 
 /* PainLevelView */
-"pain.title" = "Record pain Level";
-"pain.description" = "t your pain level";
+"pain.title" = "Record Pain Level";
+"pain.description" = "Select your pain level";
+"pain.rom_measure" = "ROM Measurement";
 
 /* Pain Categories */
 "pain.category.0" = "No pain";
@@ -105,7 +107,7 @@
 "progress.same_severepain.description" = "No change in movement, and pain continues. Have a medical specialist check it to identify the cause early.";
 "progress.same_excruciatingpain.headline" = "See a specialist promptly";
 "progress.same_excruciatingpain.description" = "Pain is severe. Don’t delay - see a medical specialist. Quick action is the first step to recovery.";
-"progress.worse_nopain.headlne" = "Movement reduced slightly";
+"progress.worse_nopain.headline" = "Movement reduced slightly";
 "progress.worse_nopain.description" = "No pain, but the joint feels a bit stiff. Move gently to loosen it, and avoid overexertion.";
 "progress.worse_mildpain.headline" = "Maybe your body feels tired";
 "progress.worse_mildpain.description" = "Range of flexion has decreased. Rest well today and check again tomorrow.";
@@ -131,7 +133,7 @@
 "history.pain_subheadline_norecord" = "Make your first record. Track your body’s changes with Anggle!";
 "history.pain_subheadline_1record" = "You’ve made your first record.Anggle cheers you on in your first step toward recovery!";
 
-"history.rom_title" = "Knee flexion range";
+"history.rom_title" = "Flexion range";
 "history.rom_semibold15_better" = "You can flex (nn)° more than previous (nnn)°";
 "history.rom_semibold15_same" = "Same as the period’s max of (nnn)°";
 "history.rom_semibold15_worse" = "You moved (nnn)° less than previous nnn°";
@@ -142,3 +144,11 @@
 "history.pain_semibold15_same" = "Same pain level as the previous lowest";
 "history.pain_semibold15_worse" = "Pain increased by (n) level from (nn)";
 "history.pain_semibold15_norecord" = "Make your first record";
+
+/* History Summary Card - ROM Change */
+"history.rom_change_increased" = "Over the past %d days,\nflexion range\nincreased by %d°!";
+"history.rom_change_decreased" = "Over the past %d days,\nflexion range\ndecreased by %d°";
+
+/* History Summary Card - Pain Change */
+"history.pain_change_decreased" = "Compared to the start,\npain level\ndecreased by %d levels!";
+"history.pain_change_increased" = "Compared to the start,\npain level\nincreased by %d levels";
